+++ Gravitas.DAL/Repository/PackingTare/IPackingTareRepository.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Linq;
 using Gravitas.Model.DomainModel.PackingTare.DTO;
 
@@ -7,11 +6,11 @@ namespace Gravitas.DAL.Repository.PackingTare
     public interface IPackingTareRepository
     {
         IQueryable<PackingTareVm> GetTareList();
-        IQueryable<TicketPackingTareVm> GetTicketTareList(long ticketId);
+        IQueryable<TicketPackingTareVm> GetTicketTareList(int ticketId);
         bool Add(PackingTareVm model);
-        void Remove(long id);
-        void RemoveTicketTare(long ticketId);
-        float GetTareWeight(long tareId);
-        bool AddTicketTare(long ticketId, long tareId, int count);
+        void Remove(int id);
+        void RemoveTicketTare(int ticketId);
+        float GetTareWeight(int tareId);
+        bool AddTicketTare(int ticketId, int tareId, int count);
     }
 }
