+++ Gravitas.DAL/Repository/PhoneInformTicketAssignment/IPhoneInformTicketAssignmentRepository.cs
@@ -5,8 +5,8 @@ namespace Gravitas.DAL.Repository.PhoneInformTicketAssignment
     public interface IPhoneInformTicketAssignmentRepository
     {
         void Add(Model.DomainModel.PhoneInformTicketAssignment.DAO.PhoneInformTicketAssignment model);
-        void Delete(long id);
+        void Delete(int id);
         IQueryable<Model.DomainModel.PhoneInformTicketAssignment.DAO.PhoneInformTicketAssignment> GetAll();
-        Model.DomainModel.PhoneInformTicketAssignment.DAO.PhoneInformTicketAssignment GetById(long id);
+        Model.DomainModel.PhoneInformTicketAssignment.DAO.PhoneInformTicketAssignment GetById(int id);
     }
 }
