+++ Gravitas.Infrastructure.Platform/Manager/LoadPoint/ILoadPointManager.cs
@@ -2,7 +2,7 @@ namespace Gravitas.Infrastructure.Platform.Manager.LoadPoint
 {
     public interface ILoadPointManager
     {
-        bool ConfirmLoadGuide(long ticketId, string employeeId);
-        bool ConfirmLoadPoint(long ticketId, string employeeId);
+        bool ConfirmLoadGuide(int ticketId, string employeeId);
+        bool ConfirmLoadPoint(int ticketId, string employeeId);
     }
 }
