+++ Gravitas.Infrastructure.Platform/Manager/Queue/Infrastructure/IQueueInfrastructure.cs
@@ -4,8 +4,8 @@ namespace Gravitas.Infrastructure.Platform.Manager.Queue.Infrastructure
 {
     public interface IQueueInfrastructure
     {
-        DateTime GetPredictionEntranceTime(long routeId);
-        void ImmediateEntrance(long ticketContainer);
-        int GetTruckCountInQueue(long routeId);
+        DateTime GetPredictionEntranceTime(int routeId);
+        void ImmediateEntrance(int ticketContainer);
+        int GetTruckCountInQueue(int routeId);
     }
 }
