+++ Gravitas.Infrastructure.Platform/Manager/UnloadPoint/IUnloadPointManager.cs
@@ -2,7 +2,7 @@ namespace Gravitas.Infrastructure.Platform.Manager.UnloadPoint
 {
     public interface IUnloadPointManager
     {
-        bool ConfirmUnloadGuide(long ticketId, string employeeId);
-        bool ConfirmUnloadPoint(long ticketId, string employeeId);
+        bool ConfirmUnloadGuide(int ticketId, string employeeId);
+        bool ConfirmUnloadPoint(int ticketId, string employeeId);
     }
 }
